pyproject.toml
[tool.poetry]
name = "bollywood-data-science"
version = "0.0.0"
description = "Quantitative analysis of the Bollywood industry"
authors = ["Samuel Grayson <sam+dev@samgrayson.me>"]
license = "NCSA"
readme = 'README.rst'
repository = "https://github.com/charmoniumQ/bollywood-data-science.git"
packages = [
    { include = "bollywood_data_science" },
]
include = ["res/"]

[tool.poetry.scripts]
# script_name = "package.module:function"

[tool.poetry.dependencies]
python = "^3.8"
rdflib = "^5.0.0"
requests = "^2.23.0"
inquirer = "^2.7.0"
pyyaml = "^5.3.1"
psutil = "^5.7.0"
"charmonium.time_block" = "^0.1.0"
"charmonium.cache" = "^0.6.0"
aiohttp = {extras = ["speedps"], version = "^3.6.2"}
aiofiles = "^0.5.0"
imdbpy = {git = "https://github.com/alberanid/imdbpy"}

[tool.poetry.dev-dependencies]
twine = "^3.1.1"
autoflake = "^1.3.1"
isort = "^4.3.21"
black = "^19.10b0"
pylint = "^2.5.2"
pylint-exit = "^1.1.0"
mypy = "^0.780"
pytest = "^5.4.1"
pytest-cov = "^2.10.0"
coverage = "^5.1"
codecov = "^2.1.4"
sphinx = "^3.1.0"
sphinx-autoapi = "^1.4.0"
sphinx-autodoc-annotation = "^1.0-1"
bump2version = "^1.0.0"
tox = "^3.15.2"
ipython = "^7.15.0"
psutil = "^5.7.0"

[tool.tox]
# According to tox docs
# The tox configuration can also be in pyproject.toml (if you want to avoid an extra file).
# Currently only the old format is supported via legacy_tox_ini, a native implementation is planned though.
legacy_tox_ini = """
[tox]
envlist = py38
isolated_build = True

[tox:.package]
# note tox will use the same python version as under what tox is installed to package
# so unless this is python 3 you can require a given python version for the packaging
# environment via the basepython key
basepython = python3

[testenv]
commands =
    poetry install
    pytest --quiet --exitfirst tests
"""
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
